moved tests to dropdown folder
